+++ CHANGES.md
@@ -1,19 +1,41 @@
-This document outlines changes since 0.3.0.  For older changelogs, see the
-[dgrid wiki](https://github.com/SitePen/dgrid/wiki).
+**Note:** Going forward, release notes will be maintained on the GitHub repository's
+[releases page](https://github.com/SitePen/dgrid/releases).  This file will likely be
+removed in the future.
 
-# master (0.3.15-dev)
+# 0.3.15
 
 ## Significant changes
 
+### General/Core
+
+* Documentation has been moved out of the wiki and into the repository under the
+  doc folder, to facilitate maintaining relevant documentation for both 0.3 and 0.4.
+* Fixed a regression in `List` which could cause errors in certain situations
+  due to a feature test being performed before the document is ready. (#907)
+
 ### Mixins
 
+* The `Keyboard` mixin no longer fires `dgrid-cellfocusout` and `dgrid-cellfocusin`
+  events when a row/cell is automatically re-focused after an observed update. (#902)
+* The `Keyboard` mixin now correctly emits `dgrid-cellfocusout` events from the
+  element losing focus, not the one gaining focus. (#917, thanks cmaus)
+* Fixed a focus issue in the `Keyboard` mixin which manifested when used in
+  conjunction with the `ColumnResizer` extension. (#928)
 * The `Selection` mixin no longer emits superfluous `dgrid-deselect` events
   for rows which were already deselected. (#889)
+* Fixed an issue in the `Selection` mixin regarding disabling text selection in
+  IE10+. (#786)
+
+### Plugins
+
+* Fixed various focus/blur-related issues in `editor` and improved tests. (#903)
 
 ### Extensions
 
 * Fixed a regression in `ColumnHider` where the node to open the menu became
   invisible on platforms with hidden scrollbars. (#886)
+* Resolved an issue regarding `dojo/dnd` CSS classes and programmatic dgrid
+  deselection. (#906)
 * `ColumnHider` now assigns its menu node an ID in the format
   `{id}-hider-menu`, not `dgrid-hider-menu-{id}`.
 
@@ -26,9 +48,18 @@ This document outlines changes since 0.3.0.  For older changelogs, see the
 
 ### Mixins
 
+* Fixed a state preservation issue in the `Keyboard` mixin when a row/cell that
+  was last focused is removed while the list/grid no longer has focus. (#899)
 * The `Selection` mixin (and `selector` column plugin) will no longer clear the
   selection during an attempt to select a row that cannot be selected
   (i.e. `allowSelect` returns `false`). (#822)
+
+### Extensions
+
+* Fixed an IE10 issue in `ColumnHider` where header cells (but not body cells)
+  would remain hidden after columns have been resized with `ColumnResizer`. (#841)
+* Fixed an issue in `CompoundColumns` where sorting programmatically would
+  cause an error when attempting to place the sort arrow node. (#901)
 
 # 0.3.14